acrn: add feature to patch configuration
we can now generate the configuration and patch it before build it.
As this is not patching the original source code it does not work with
SRC_URI, so add a new variable ACRN_CONFIG_PATCH to point to the patch file.

add it to local.conf to apply the patch to the configuration:
ACRN_CONFIG_PATCH = "/home/user/config.patch"

reference:
https://projectacrn.github.io/latest/tutorials/acrn_configuration_tool.html#makefile-targets-for-configuration

Signed-off-by: Lee Chee Yang <chee.yang.lee@intel.com>